+++ routes/api.go
@@ -17,7 +17,4 @@ func API(app *core.Application) {
 			})
 		})
 	})
-
-	RegisterAuthAPI(app)
-	RegisterDashboardAPI(app)
 }